Move submodules into folder, establish utils.py
also renamed alreadyDrawnOps
