+++ damagecalc/damage_formulas.py
